Реализовал жанры фильмов во фрагменте синхронно
