+++ app/src/main/java/ru/dudar/findfilms/domain/TheMovieGenres.kt
@@ -1,9 +1,9 @@
-package ru.dudar.findfilms.data
+package ru.dudar.findfilms.domain
 
 import com.google.gson.annotations.SerializedName
 
 
-data class Example (
+data class TheMovieGenres (
 
     @SerializedName("genres") var genres : ArrayList<Genres>
