+++ app/src/main/java/ru/dudar/findfilms/domain/Themoviedb.kt
@@ -1,4 +1,10 @@
 package ru.dudar.findfilms.domain
 
+import androidx.annotation.WorkerThread
+
 interface Themoviedb {
+    @WorkerThread
+    fun getTMDBGenresSync(): List<TheMovieGenres>
+
+    fun getTMDBGenresAsync(callback: (List<TheMovieGenres>) -> Unit)
 }
